+++ routes/apiRoutes.js
@@ -1,4 +1,4 @@
-var db = require("../models");
+var db = require("../models/indexModel");
 
 module.exports = function(app) {
   // Get all accounts
